requirements.txt
pyaudio
openai-whisper
numpy